requirements.txt
aiofiles==24.1.0
altair==5.4.1
analytics-python==1.4.post1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==1.10.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.0.9
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-grappelli==4.0.1
django-import-export==4.2.0
django-js-asset==2.2.0
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.20.0
filelock==3.16.1
fitz==0.0.1.dev2
fqdn==1.5.1
frontend==0.0.3
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.13.3
idna==3.10
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.29.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mitosheet==0.2.2
monotonic==1.6
mysql-connector-python==9.1.0
# mysqlclient==2.2.5
narwhals==1.12.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.8.6
notebook_shim==0.2.4
numpy==2.1.2
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib==1.0.1
pillow==10.4.0
pipdeptree==2.23.4
platformdirs==4.3.6
plotly==5.24.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
prov==2.0.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydot==3.0.2
Pygments==2.18.0
PyMuPDF==1.24.13
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
referencing==0.35.1
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rpds-py==0.20.0
rsa==4.9
scipy==1.14.1
Send2Trash==1.8.3
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.41.2
streamlit==1.39.0
streamlit-option-menu==0.4.0
tablib==3.7.0
tenacity==9.0.0
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
tomli==2.0.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
traits==6.3.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
watchdog==5.0.3
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
mysql_connector_repackaged==0.3.1
openai==0.28.1
python_docx==1.1.2
scikit_learn==1.5.2
streamlit_ext==0.1.12
streamlit_option_menu==0.4.0
torch==2.5.1
transformers==4.46.1
cryptography==43.0.3
dnspython==2.7.0
googletrans==3.0.0